+++ Api/Infra/Repository/SalaryRepository.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq.Expressions;
 using System.Threading.Tasks;
-using Api.Infra.Resources.Salary;
+using Cashflow.Api.Infra.Sql.Salary;
 using Cashflow.Api.Infra.Entity;
 using Cashflow.Api.Service;
 using Cashflow.Api.Shared;
@@ -17,13 +17,13 @@ namespace Cashflow.Api.Infra.Repository
 
         public Task<IEnumerable<Salary>> GetAll() => throw new NotImplementedException();
 
-        public Task<Salary> GetById(int id) => FirstOrDefault(SalaryResources.ById, new { Id = id });
+        public Task<Salary> GetById(long id) => FirstOrDefault(SalaryResources.ById, new { Id = id });
 
         public Task<IEnumerable<Salary>> GetByUserId(int userId) => Query(SalaryResources.ByUser, new { UserId = userId });
 
         public Task<IEnumerable<Salary>> GetSome(Expression<Func<Salary, bool>> expressions) => throw new NotImplementedException();
 
-        public Task Remove(int id) => Execute(SalaryResources.Delete, new { Id = id });
+        public Task Remove(long id) => Execute(SalaryResources.Delete, new { Id = id });
 
         public Task Update(Salary salary) => Execute(SalaryResources.Update, salary);
     }
