+++ Api/Infra/Repository/UserRepository.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Cashflow.Api.Infra.Entity;
-using Cashflow.Api.Infra.Resources.User;
+using Cashflow.Api.Infra.Sql.User;
 using Cashflow.Api.Service;
 using Cashflow.Api.Shared;
 
@@ -14,7 +14,7 @@ namespace Cashflow.Api.Infra.Repository
 
         public Task<User> FindByNickName(string nickName) => FirstOrDefault(UserResources.ByNickName, new { NickName = nickName });
 
-        public Task<User> GetById(int id) => FirstOrDefault(UserResources.ById, new { Id = id });
+        public Task<User> GetById(long id) => FirstOrDefault(UserResources.ById, new { Id = id });
 
         public async Task<IEnumerable<User>> GetAll() => await Query(UserResources.All);
 
@@ -22,6 +22,6 @@ namespace Cashflow.Api.Infra.Repository
 
         public Task Update(User t) => throw new NotImplementedException();
 
-        public Task Remove(int id) => throw new NotImplementedException();
+        public Task Remove(long id) => throw new NotImplementedException();
     }
 }
